+++ include/Game/Entities/ItemBridge.h
@@ -160,7 +160,7 @@ struct Mgr : public TNodeItemMgr {
 	virtual u32 generatorGetID() { return 'brdg'; }                       // _58 (weak)
 	virtual BaseItem* generatorBirth(Vector3f&, Vector3f&, GenItemParm*); // _5C
 	virtual void generatorWrite(Stream&, GenItemParm*);                   // _60
-	virtual void generatorRead(Stream&, GenItemParm*, unsigned long);     // _64
+	virtual void generatorRead(Stream&, GenItemParm*, u32);               // _64
 	virtual u32 generatorLocalVersion() { return '0001'; }                // _68 (weak)
 	virtual GenItemParm* generatorNewItemParm();                          // _70
 	virtual BaseItem* birth();                                            // _BC
