+++ include/Game/Entities/ItemCave.h
@@ -141,7 +141,7 @@ struct Mgr : public TNodeItemMgr {
 	virtual u32 generatorGetID() { return 'cave'; }                       // _58 (weak)
 	virtual BaseItem* generatorBirth(Vector3f&, Vector3f&, GenItemParm*); // _5C
 	virtual void generatorWrite(Stream&, GenItemParm*);                   // _60
-	virtual void generatorRead(Stream&, GenItemParm*, unsigned long);     // _64
+	virtual void generatorRead(Stream&, GenItemParm*, u32);               // _64
 	virtual u32 generatorLocalVersion() { return '0002'; }                // _68 (weak)
 	virtual GenItemParm* generatorNewItemParm();                          // _70
 	virtual BaseItem* doNew() { return new Item; }                        // _A0 (weak)
