+++ include/Game/Entities/ItemDengekiGate.h
@@ -12,7 +12,7 @@ struct Mgr : public NodeItemMgr<ItemGate> {
 	virtual u32 generatorGetID();                                         // _58 (weak)
 	virtual BaseItem* generatorBirth(Vector3f&, Vector3f&, GenItemParm*); // _5C
 	virtual void generatorWrite(Stream&, GenItemParm*);                   // _60
-	virtual void generatorRead(Stream&, GenItemParm*, unsigned long);     // _64
+	virtual void generatorRead(Stream&, GenItemParm*, u32);               // _64
 	virtual u32 generatorLocalVersion();                                  // _68 (weak)
 	virtual GenItemParm* generatorNewItemParm();                          // _70
 	virtual ~Mgr();                                                       // _B4 (weak)
