+++ include/JSystem/J3D/J3DAnmLoader.h
@@ -23,7 +23,7 @@ struct J3DAnmLoaderDataBase {
 enum J3DMtxCalcFlag { MTXCalc_0 };
 
 J3DMtxCalcAnmBase* J3DNewMtxCalcAnm(u32, J3DAnmTransform*);
-J3DMtxCalc* J3DUNewMtxCalcAnm(unsigned long, J3DAnmTransform*, J3DAnmTransform*, J3DAnmTransform*, J3DAnmTransform*, J3DMtxCalcFlag);
+J3DMtxCalc* J3DUNewMtxCalcAnm(u32, J3DAnmTransform*, J3DAnmTransform*, J3DAnmTransform*, J3DAnmTransform*, J3DMtxCalcFlag);
 
 struct J3DAnmLoader {
 	virtual J3DAnmBase* load(const void*)              = 0; // _08
