+++ include/JSystem/JStage/TSystem.h
@@ -9,10 +9,10 @@ struct TSystem : public TObject {
 	virtual ~TSystem() = 0;                                            // _08
 	virtual int JSGFGetType() const;                                   // _0C
 	virtual int JSGFindObject(TObject**, const char*, TEObject) const; // _3C
-	virtual int JSGCreateObject(const char*, TEObject, unsigned long); // _40
+	virtual int JSGCreateObject(const char*, TEObject, u32);           // _40
 	virtual void JSGDestroyObject(TObject*);                           // _44
-	virtual int JSGGetSystemData(unsigned long);                       // _48
-	virtual void JSGSetSystemData(unsigned long, unsigned long);       // _4C
+	virtual int JSGGetSystemData(u32);                                 // _48
+	virtual void JSGSetSystemData(u32, u32);                           // _4C
 };
 } // namespace JStage
