+++ include/JSystem/JStudio/TAdaptor.h
@@ -72,19 +72,19 @@ struct TAdaptor_actor : public TAdaptor {
 
 	// ~TAdaptor_actor();
 
-	virtual ~TAdaptor_actor()                                                                                  = 0; // _08
-	virtual void adaptor_do_PARENT(JStudio::data::TEOperationData, const void*, unsigned long)                 = 0; // _20
-	virtual void adaptor_do_PARENT_NODE(JStudio::data::TEOperationData, const void*, unsigned long)            = 0; // _24
-	virtual void adaptor_do_PARENT_ENABLE(JStudio::data::TEOperationData, const void*, unsigned long)          = 0; // _28
-	virtual void adaptor_do_PARENT_FUNCTION(JStudio::data::TEOperationData, const void*, unsigned long)        = 0; // _2C
-	virtual void adaptor_do_RELATION(JStudio::data::TEOperationData, const void*, unsigned long)               = 0; // _30
-	virtual void adaptor_do_RELATION_NODE(JStudio::data::TEOperationData, const void*, unsigned long)          = 0; // _34
-	virtual void adaptor_do_RELATION_ENABLE(JStudio::data::TEOperationData, const void*, unsigned long)        = 0; // _38
-	virtual void adaptor_do_SHAPE(JStudio::data::TEOperationData, const void*, unsigned long)                  = 0; // _3C
-	virtual void adaptor_do_ANIMATION(JStudio::data::TEOperationData, const void*, unsigned long)              = 0; // _40
-	virtual void adaptor_do_ANIMATION_MODE(JStudio::data::TEOperationData, const void*, unsigned long)         = 0; // _44
-	virtual void adaptor_do_TEXTURE_ANIMATION(JStudio::data::TEOperationData, const void*, unsigned long)      = 0; // _48
-	virtual void adaptor_do_TEXTURE_ANIMATION_MODE(JStudio::data::TEOperationData, const void*, unsigned long) = 0; // _4C
+	virtual ~TAdaptor_actor()                                                                        = 0; // _08
+	virtual void adaptor_do_PARENT(JStudio::data::TEOperationData, const void*, u32)                 = 0; // _20
+	virtual void adaptor_do_PARENT_NODE(JStudio::data::TEOperationData, const void*, u32)            = 0; // _24
+	virtual void adaptor_do_PARENT_ENABLE(JStudio::data::TEOperationData, const void*, u32)          = 0; // _28
+	virtual void adaptor_do_PARENT_FUNCTION(JStudio::data::TEOperationData, const void*, u32)        = 0; // _2C
+	virtual void adaptor_do_RELATION(JStudio::data::TEOperationData, const void*, u32)               = 0; // _30
+	virtual void adaptor_do_RELATION_NODE(JStudio::data::TEOperationData, const void*, u32)          = 0; // _34
+	virtual void adaptor_do_RELATION_ENABLE(JStudio::data::TEOperationData, const void*, u32)        = 0; // _38
+	virtual void adaptor_do_SHAPE(JStudio::data::TEOperationData, const void*, u32)                  = 0; // _3C
+	virtual void adaptor_do_ANIMATION(JStudio::data::TEOperationData, const void*, u32)              = 0; // _40
+	virtual void adaptor_do_ANIMATION_MODE(JStudio::data::TEOperationData, const void*, u32)         = 0; // _44
+	virtual void adaptor_do_TEXTURE_ANIMATION(JStudio::data::TEOperationData, const void*, u32)      = 0; // _48
+	virtual void adaptor_do_TEXTURE_ANIMATION_MODE(JStudio::data::TEOperationData, const void*, u32) = 0; // _4C
 
 	JStudio::TVariableValue _0C[14]; // _0C
 
@@ -119,14 +119,14 @@ struct TAdaptor_camera : public TAdaptor {
 
 	// ~TAdaptor_camera();
 
-	virtual ~TAdaptor_camera()                                                                               = 0; // _08
-	virtual void adaptor_do_PARENT(JStudio::data::TEOperationData, const void*, unsigned long)               = 0; // _20
-	virtual void adaptor_do_PARENT_NODE(JStudio::data::TEOperationData, const void*, unsigned long)          = 0; // _24
-	virtual void adaptor_do_PARENT_ENABLE(JStudio::data::TEOperationData, const void*, unsigned long)        = 0; // _28
-	virtual void adaptor_do_PARENT_FUNCTION(JStudio::data::TEOperationData, const void*, unsigned long)      = 0; // _2C
-	virtual void adaptor_do_TARGET_PARENT(JStudio::data::TEOperationData, const void*, unsigned long)        = 0; // _30
-	virtual void adaptor_do_TARGET_PARENT_NODE(JStudio::data::TEOperationData, const void*, unsigned long)   = 0; // _34
-	virtual void adaptor_do_TARGET_PARENT_ENABLE(JStudio::data::TEOperationData, const void*, unsigned long) = 0; // _38
+	virtual ~TAdaptor_camera()                                                                     = 0; // _08
+	virtual void adaptor_do_PARENT(JStudio::data::TEOperationData, const void*, u32)               = 0; // _20
+	virtual void adaptor_do_PARENT_NODE(JStudio::data::TEOperationData, const void*, u32)          = 0; // _24
+	virtual void adaptor_do_PARENT_ENABLE(JStudio::data::TEOperationData, const void*, u32)        = 0; // _28
+	virtual void adaptor_do_PARENT_FUNCTION(JStudio::data::TEOperationData, const void*, u32)      = 0; // _2C
+	virtual void adaptor_do_TARGET_PARENT(JStudio::data::TEOperationData, const void*, u32)        = 0; // _30
+	virtual void adaptor_do_TARGET_PARENT_NODE(JStudio::data::TEOperationData, const void*, u32)   = 0; // _34
+	virtual void adaptor_do_TARGET_PARENT_ENABLE(JStudio::data::TEOperationData, const void*, u32) = 0; // _38
 
 	JStudio::TVariableValue _0C[0xC]; // _0C
 
@@ -158,9 +158,9 @@ struct TAdaptor_light : public TAdaptor {
 	}
 	// ~TAdaptor_light();
 
-	virtual ~TAdaptor_light()                                                                   = 0; // _08
-	virtual void adaptor_do_ENABLE(JStudio::data::TEOperationData, const void*, unsigned long)  = 0; // _20
-	virtual void adaptor_do_FACULTY(JStudio::data::TEOperationData, const void*, unsigned long) = 0; // _24
+	virtual ~TAdaptor_light()                                                         = 0; // _08
+	virtual void adaptor_do_ENABLE(JStudio::data::TEOperationData, const void*, u32)  = 0; // _20
+	virtual void adaptor_do_FACULTY(JStudio::data::TEOperationData, const void*, u32) = 0; // _24
 
 	static const u32 sauVariableValue_3_COLOR_RGB[3];
 	static const u32 sauVariableValue_4_COLOR_RGBA[4];
@@ -189,15 +189,15 @@ struct TAdaptor_particle : public TAdaptor {
 
 	// ~TAdaptor_particle();
 
-	virtual ~TAdaptor_particle()                                                                      = 0; // _08
-	virtual void adaptor_do_PARTICLE(JStudio::data::TEOperationData, const void*, unsigned long)      = 0; // _20
-	virtual void adaptor_do_BEGIN(JStudio::data::TEOperationData, const void*, unsigned long)         = 0; // _24
-	virtual void adaptor_do_BEGIN_FADE_IN(JStudio::data::TEOperationData, const void*, unsigned long) = 0; // _28
-	virtual void adaptor_do_END(JStudio::data::TEOperationData, const void*, unsigned long)           = 0; // _2C
-	virtual void adaptor_do_END_FADE_OUT(JStudio::data::TEOperationData, const void*, unsigned long)  = 0; // _30
-	virtual void adaptor_do_PARENT(JStudio::data::TEOperationData, const void*, unsigned long)        = 0; // _34
-	virtual void adaptor_do_PARENT_NODE(JStudio::data::TEOperationData, const void*, unsigned long)   = 0; // _38
-	virtual void adaptor_do_PARENT_ENABLE(JStudio::data::TEOperationData, const void*, unsigned long) = 0; // _3C
+	virtual ~TAdaptor_particle()                                                            = 0; // _08
+	virtual void adaptor_do_PARTICLE(JStudio::data::TEOperationData, const void*, u32)      = 0; // _20
+	virtual void adaptor_do_BEGIN(JStudio::data::TEOperationData, const void*, u32)         = 0; // _24
+	virtual void adaptor_do_BEGIN_FADE_IN(JStudio::data::TEOperationData, const void*, u32) = 0; // _28
+	virtual void adaptor_do_END(JStudio::data::TEOperationData, const void*, u32)           = 0; // _2C
+	virtual void adaptor_do_END_FADE_OUT(JStudio::data::TEOperationData, const void*, u32)  = 0; // _30
+	virtual void adaptor_do_PARENT(JStudio::data::TEOperationData, const void*, u32)        = 0; // _34
+	virtual void adaptor_do_PARENT_NODE(JStudio::data::TEOperationData, const void*, u32)   = 0; // _38
+	virtual void adaptor_do_PARENT_ENABLE(JStudio::data::TEOperationData, const void*, u32) = 0; // _3C
 
 	JStudio::TVariableValue _0C[18];
