+++ include/JSystem/JStudio/TParse.h
@@ -8,11 +8,11 @@
 namespace JStudio {
 struct TParse : public stb::TParse {
 	TParse(JStudio::TControl*);
-	virtual ~TParse();                                                             // _08
-	virtual bool parseHeader(const stb::data::TParse_THeader&, unsigned long);     // _14
-	virtual void parseBlock_block(const stb::data::TParse_TBlock&, unsigned long); // _18
+	virtual ~TParse();                                                   // _08
+	virtual bool parseHeader(const stb::data::TParse_THeader&, u32);     // _14
+	virtual void parseBlock_block(const stb::data::TParse_TBlock&, u32); // _18
 
-	void parseBlock_block_fvb_(const JStudio::stb::data::TParse_TBlock&, unsigned long);
+	void parseBlock_block_fvb_(const JStudio::stb::data::TParse_TBlock&, u32);
 
 	fvb::TParse mFvbParse; // _08
 };
