+++ include/JSystem/JStudio_JMessage.h
@@ -25,8 +25,8 @@ struct TCreateObject : public TCreateObject_base {
 };
 
 struct TAdaptor_message : public JStudio::TAdaptor_message {
-	virtual ~TAdaptor_message();                                                                 // _08
-	virtual void adaptor_do_MESSAGE(JStudio::data::TEOperationData, const void*, unsigned long); // _20
+	virtual ~TAdaptor_message();                                                       // _08
+	virtual void adaptor_do_MESSAGE(JStudio::data::TEOperationData, const void*, u32); // _20
 
 	TAdaptor_message(JMessage::TControl*);
 	JMessage::TControl* mControl; // _10
