+++ include/JSystem/JUtility/JUTVideo.h
@@ -20,8 +20,8 @@ struct JUTVideo {
 	static JUTVideo* createManager(const GXRenderModeObj*);
 	static void destroyManager();
 	static JUTVideo* getManager() { return sManager; }
-	static void preRetraceProc(unsigned long);
-	static void postRetraceProc(unsigned long);
+	static void preRetraceProc(u32);
+	static void postRetraceProc(u32);
 	static void drawDoneCallback();
 
 	inline static u32 getVideoInterval() { return sVideoInterval; }
